LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-serialization/cssStyleValue-cssom-expected.txt


PASS CSSStyleValue from specified CSSOM serializes correctly
FAIL CSSStyleValue from computed CSSOM serializes correctly elem.computedStyleMap is not a function. (In 'elem.computedStyleMap()', 'elem.computedStyleMap' is undefined)
FAIL Shorthand CSSStyleValue from inline CSSOM serializes correctly assert_equals: expected "blue" but got "initial"
FAIL Shorthand CSSStyleValue from computed CSSOM serializes correctly elem.computedStyleMap is not a function. (In 'elem.computedStyleMap()', 'elem.computedStyleMap' is undefined)
